Module Dns_client

TODO ideally there'd be something like mirage-flow-lwt that didn't depend on lwt and a ton of other things, and still provided map and connect and so on. leaving this stuff here for now until a better solution presents itself.

val default_resolver : string

default_resolver is the IPv4 address in dotted-decimal form of the default resolver. Currently it is the IP address of the UncensoredDNS.org anycast service.

module type S = sig ... end

module Make : functor (T : S) -> sig ... end

module Pure : sig ... end