Module 4-S.TCPV4

type error#row

The type for TCP errors.

and error = private [>

| Mirage_protocols.Tcp.error

]

The type for TCP errors.

type write_error#row

The type for TCP write errors.

and write_error = private [>

| Mirage_protocols.Tcp.write_error

]

The type for TCP write errors.

type ipaddr = Ipaddr.V4.t

The type for IP address representations.

type ipinput

The type for input function continuation to pass onto the underlying IP layer. This will normally be a NOOP for a conventional kernel, but a direct implementation will parse the buffer.

type flow

A flow represents the state of a single TCPv4 stream that is connected to an endpoint.

include Mirage_device.S

type t

The type representing the internal state of the device

val disconnect : t -> unit Lwt.t

Disconnect from the device. While this might take some time to complete, it can never result in an error.

include Mirage_flow.S with type flow := flow and type error := error and type write_error := write_error

val pp_error : error Fmt.t

pp_error is the pretty-printer for errors.

type nonrec write_error#row

The type for write errors.

val pp_write_error : write_error Fmt.t

pp_write_error is the pretty-printer for write errors.

val read : flow -> (Cstruct.t Mirage_flow.or_eof, error) Stdlib.result Lwt.t

read flow blocks until some data is available and returns a fresh buffer containing it.

The returned buffer will be of a size convenient to the flow implementation, but will always have at least 1 byte.

If the remote endpoint calls close then calls to read will keep returning data until all the in-flight data has been read. read flow will return `Eof when the remote endpoint has called close and when there is no more in-flight data.

val write : flow -> Cstruct.t -> (unit, write_error) Stdlib.result Lwt.t

write flow buffer writes a buffer to the flow. There is no indication when the buffer has actually been read and, therefore, it must not be reused. The contents may be transmitted in separate packets, depending on the underlying transport. The result Ok () indicates success, Error `Closed indicates that the connection is now closed and therefore the data could not be written. Other errors are possible.

val writev : flow -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

writev flow buffers writes a sequence of buffers to the flow. There is no indication when the buffers have actually been read and, therefore, they must not be reused. The result Ok () indicates success, Error `Closed indicates that the connection is now closed and therefore the data could not be written. Other errors are possible.

val close : flow -> unit Lwt.t

close flow flushes all pending writes and signals the remote endpoint that there will be no future writes. Once the remote endpoint has read all pending data, it is expected that calls to read on the remote return `Eof.

Note it is still possible for the remote endpoint to write to the flow and for the local endpoint to call read. This state where the local endpoint has called close but the remote endpoint has not called close is similar to that of a half-closed TCP connection or a Unix socket after shutdown(SHUTDOWN_WRITE).

close flow waits until the remote endpoint has also called close before returning. At this point no data can flow in either direction and resources associated with the flow can be freed.

val dst : flow -> ipaddr * int

Get the destination IPv4 address and destination port that a flow is currently connected to.

val write_nodelay : flow -> Cstruct.t -> (unit, write_error) Stdlib.result Lwt.t

write_nodelay flow buffer writes the contents of buffer to the flow. The thread blocks until all data has been successfully transmitted to the remote endpoint. Buffering within the layer is minimized in this mode. Note that this API will change in a future revision to be a per-flow attribute instead of a separately exposed function.

val writev_nodelay : flow -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

writev_nodelay flow buffers writes the contents of buffers to the flow. The thread blocks until all data has been successfully transmitted to the remote endpoint. Buffering within the layer is minimized in this mode. Note that this API will change in a future revision to be a per-flow attribute instead of a separately exposed function.

val create_connection : ?⁠keepalive:Mirage_protocols.Keepalive.t -> t -> (ipaddr * int) -> (flow, error) Stdlib.result Lwt.t

create_connection ~keepalive t (addr,port) opens a TCPv4 connection to the specified endpoint.

If the optional argument ?keepalive is provided then TCP keep-alive messages will be sent to the server when the connection is idle. If no responses are received then eventually the connection will be disconnected: read will return Ok `Eof and write will return Error `Closed

type listener = {

process : flow -> unit Lwt.t;
process a connected flow

keepalive : Mirage_protocols.Keepalive.t option;
optional TCP keepalive configuration

}

A TCP listener on a particular port

val input : t -> listeners:(int -> listener option) -> ipinput

input t listeners returns an input function continuation to be passed to the underlying IP layer.

When the layer receives a TCP SYN (i.e. a connection request) to a particular port, it will evaluate listeners port:

If listeners port is None, the input function will return an RST to refuse the connection.

If listeners port is Some listener then the connection will be accepted and the resulting flow will be processed by listener.process. If listener.keepalive is Some configuration then the TCP keep-alive configuration will be applied before calling listener.process.