Module Dns_client

TODO ideally there'd be something like mirage-flow-lwt that didn't depend on lwt and a ton of other things, and still provided map and connect and so on. leaving this stuff here for now until a better solution presents itself.

val stdlib_random : int -> Cstruct.t

stdlib_random len is a buffer of size len, filled with random data. This function is used by default (in the Unix and Lwt implementations) for filling the ID field of the DNS packet. Internally, the Random module from the OCaml standard library is used, which is not cryptographically secure. If desired Nocrypto.Rng.generate can be passed to S.create.

module type S = sig ... end

module Make : functor (T : S) -> sig ... end

module Pure : sig ... end